feat!: Combine GlobStream & GlobReadable into one API
